Add files via upload
All source files are uploaded directly.  Some make use of a resource fork and this is lost when moving through modern OSes so a .sit archive with all of the project files is also included.  Theoretically the files in the .sit should match what's outside the .sit but trust the .sit file contents if you want to download this and build it on your vintage Macintosh environment.  Refer to all the other files in so much as non-resource-fork portions can be version tracked still (e.g. the .c and .h files' materials should all be there and you can use git's standard versioning tools to monitor and track those files' contents).